backend/requirements.txt
fastapi
uvicorn
python-dotenv
httpx
pydantic
black
flake8
isort
pre-commit

requests
sqlalchemy
websocket-client
